fix bigints

+++ database/migrations/2022_06_27_131828_create_folio_authentication_tokens_table.php
@@ -15,7 +15,7 @@ class CreateFolioAuthenticationTokensTable extends Migration
     {
         Schema::create('folio_authentication_tokens', function (Blueprint $table) {
             $table->bigIncrements('id');
-            $table->bigInteger('institution_id');
+            $table->integer('institution_id');
             $table->string('institution_name', 50);
             $table->string('auth_key', 300)->unique('auth_key');
             $table->string('tenant', 50);
