+++ database/migrations/2022_06_27_131844_create_shelves_table.php
@@ -15,7 +15,7 @@ class CreateShelvesTable extends Migration
     {
         Schema::create('shelves', function (Blueprint $table) {
             $table->bigIncrements('id');
-            $table->bigInteger('scan_order');
+            $table->integer('scan_order');
             $table->bigInteger('user_id');
             $table->string('callnumber', 250);
             $table->string('barcode', 250);
