+++ database/migrations/2022_10_12_230301_create_corrections_table.php
@@ -15,8 +15,8 @@ return new class extends Migration
     {
         Schema::create('corrections', function (Blueprint $table) {
             $table->bigIncrements('id'); // Auto incrementing primary key
-            $table->bigInteger('user_id');
-            $table->bigInteger('library_id');
+            $table->integer('user_id');
+            $table->integer('library_id');
             $table->string('barcode', 30)->nullable();
             $table->string('call_number')->nullable();
             $table->text('title')->nullable();
